+++ bugkillers-core/src/main/java/org/bugkillers/core/domain/TagDO.java
@@ -26,7 +26,7 @@ public class TagDO {
     /**
      *  1=正常  2=删除,所属表字段为bk_tag.dr
      */
-    private Byte dr;
+    private Integer dr;
 
     /**
      * 获取 主键 字段:bk_tag.id
@@ -105,7 +105,7 @@ public class TagDO {
      *
      * @return bk_tag.dr, 1=正常  2=删除
      */
-    public Byte getDr() {
+    public Integer getDr() {
         return dr;
     }
 
@@ -114,7 +114,7 @@ public class TagDO {
      *
      * @param dr bk_tag.dr, 1=正常  2=删除
      */
-    public void setDr(Byte dr) {
+    public void setDr(Integer dr) {
         this.dr = dr;
     }
 }
